split Config.hpp
